Update scripts, secure nginx, add README.md

+++ consul-template/send_signal.sh
@@ -1,5 +1,6 @@
-#!/bin/sh
+#!/bin/sh -e
 container_id=$(/docker/docker ps --filter "name=$1" --format "{{.ID}}")
-echo "killing SIGHUP $container_id"
-/docker/docker kill --signal=SIGHUP $container_id
-exit $?
+if [ -n "$container_id" ]; then
+    echo "killing SIGHUP $container_id"
+    /docker/docker kill --signal=SIGHUP $container_id
+fi
